# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RaktFlow API",
    "description": "Verified blood donation and emergency logistics orchestration API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/bootstrap": {
      "post": {
        "tags": [
          "accounts and donor profile"
        ],
        "summary": "Bootstrap Account",
        "operationId": "bootstrap_account_api_v1_auth_bootstrap_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-uid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Uid"
            }
          },
          {
            "name": "x-dev-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/donors/me/profile": {
      "get": {
        "tags": [
          "accounts and donor profile"
        ],
        "summary": "Get Donor Profile",
        "operationId": "get_donor_profile_api_v1_donors_me_profile_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-uid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Uid"
            }
          },
          {
            "name": "x-dev-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonorProfileView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "accounts and donor profile"
        ],
        "summary": "Upsert Donor Profile",
        "operationId": "upsert_donor_profile_api_v1_donors_me_profile_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-uid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Uid"
            }
          },
          {
            "name": "x-dev-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DonorProfileUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DonorProfileView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/donors/me/screenings": {
      "post": {
        "tags": [
          "accounts and donor profile"
        ],
        "summary": "Submit Screening",
        "operationId": "submit_screening_api_v1_donors_me_screenings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-dev-uid",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Uid"
            }
          },
          {
            "name": "x-dev-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreeningSubmission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScreeningResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BootstrapResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "needs_profile": {
            "type": "boolean",
            "title": "Needs Profile"
          },
          "token_refresh_required": {
            "type": "boolean",
            "title": "Token Refresh Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "roles",
          "needs_profile"
        ],
        "title": "BootstrapResponse"
      },
...
      "DonorProfileUpsert": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Full Name"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "title": "Date Of Birth"
          },
          "phone": {
            "type": "string",
            "maxLength": 24,
            "minLength": 10,
            "title": "Phone"
          },
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "City"
          },
          "blood_type": {
            "type": "string",
            "enum": [
              "UNKNOWN",
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-",
              "BOMBAY"
            ],
            "title": "Blood Type",
            "default": "UNKNOWN"
          },
          "consent_to_process": {
            "type": "boolean",
            "title": "Consent To Process"
          },
          "emergency_notifications": {
            "type": "boolean",
            "title": "Emergency Notifications",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "date_of_birth",
          "phone",
          "city",
          "consent_to_process"
        ],
        "title": "DonorProfileUpsert"
      },
      "DonorProfileView": {
        "properties": {
          "reference_code": {
            "type": "string",
            "title": "Reference Code"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "phone_masked": {
            "type": "string",
            "title": "Phone Masked"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "blood_type": {
            "type": "string",
            "title": "Blood Type"
          },
          "profile_status": {
            "type": "string",
            "title": "Profile Status"
          },
          "email_verified": {
            "type": "boolean",
            "title": "Email Verified",
            "default": true
          },
          "identity_verified": {
            "type": "boolean",
            "title": "Identity Verified"
          },
          "latest_screening_outcome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Screening Outcome"
          }
        },
        "type": "object",
        "required": [
          "reference_code",
          "full_name",
          "age",
          "phone_masked",
          "city",
          "blood_type",
          "profile_status",
          "identity_verified",
          "latest_screening_outcome"
        ],
        "title": "DonorProfileView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScreeningResult": {
        "properties": {
          "screening_id": {
            "type": "string",
            "format": "uuid",
            "title": "Screening Id"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "PROCEED_TO_CLINICAL",
              "CLINICAL_REVIEW",
              "TEMPORARY_DEFERRAL_SUGGESTED"
            ],
            "title": "Outcome"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "valid_until": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Until"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "screening_id",
          "outcome",
          "flags",
          "valid_until",
          "message"
        ],
        "title": "ScreeningResult"
      },
      "ScreeningSubmission": {
        "properties": {
          "questionnaire_version": {
            "type": "string",
            "const": "IN-PRECHECK-2026-01",
            "title": "Questionnaire Version",
            "default": "IN-PRECHECK-2026-01"
          },
          "weight_kg": {
            "type": "number",
            "maximum": 250.0,
            "minimum": 25.0,
            "title": "Weight Kg"
          },
          "feeling_well_today": {
            "type": "boolean",
            "title": "Feeling Well Today"
          },
          "fever_infection_or_antibiotics": {
            "type": "boolean",
            "title": "Fever Infection Or Antibiotics"
          },
          "medication_requires_review": {
            "type": "boolean",
            "title": "Medication Requires Review"
          },
          "heart_lung_kidney_liver_or_bleeding_condition": {
            "type": "boolean",
            "title": "Heart Lung Kidney Liver Or Bleeding Condition"
          },
          "surgery_transfusion_or_hospitalization_last_12_months": {
            "type": "boolean",
            "title": "Surgery Transfusion Or Hospitalization Last 12 Months"
          },
          "tattoo_or_piercing_last_12_months": {
            "type": "boolean",
            "title": "Tattoo Or Piercing Last 12 Months"
          },
          "malaria_risk_travel_or_residence": {
            "type": "boolean",
            "title": "Malaria Risk Travel Or Residence"
          },
          "pregnancy_breastfeeding_or_recent_delivery": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pregnancy Breastfeeding Or Recent Delivery"
          },
          "last_donation_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Donation Date"
          },
          "answers_are_truthful": {
            "type": "boolean",
            "title": "Answers Are Truthful"
          },
          "consent_to_clinical_review": {
            "type": "boolean",
            "title": "Consent To Clinical Review"
          }
        },
        "type": "object",
        "required": [
          "weight_kg",
          "feeling_well_today",
          "fever_infection_or_antibiotics",
          "medication_requires_review",
          "heart_lung_kidney_liver_or_bleeding_condition",
          "surgery_transfusion_or_hospitalization_last_12_months",
          "tattoo_or_piercing_last_12_months",
          "malaria_risk_travel_or_residence",
          "answers_are_truthful",
          "consent_to_clinical_review"
        ],
        "title": "ScreeningSubmission"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}